pyproject.toml
[tool.poetry]
name = "mwaa-airflow"
version = "1"
description = "Pacote que faz deploy do serviço MWAA via cdk python"
authors = ["GabrielMeelo97 <gabrielmelo_@outlook.com>"]

[tool.poetry.dependencies]
python = "^3.8"
"aws-cdk.aws-ec2" = "^1.159.0"
"aws-cdk.aws-iam" = "^1.159.0"
"aws-cdk.aws-s3" = "^1.159.0"
"aws-cdk.aws-s3-deployment" = "^1.159.0"
"aws-cdk.aws-mwaa" = "^1.159.0"
"aws-cdk.core" = "^1.159.0"
"aws-cdk.aws-kms" = "^1.159.0"

[tool.poetry.dev-dependencies]

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"
